% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggravated_assaults.R
\docType{data}
\name{aggravated_assaults}
\alias{aggravated_assaults}
\title{Aggravated assaults in Austin in 2019}
\format{
A data frame containing 8,696 rows and 5 columns:
\describe{
\item{date}{A date-time}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
\item{location_type}{Type of place}
\item{location_category}{More-detailed type of place}
}

A version of this dataset is also available as a stand-alone file at
\url{https://mpjashby.github.io/crimemappingdata/aggravated_assaults.xlsx} but
the contents of the stand-alone file differ from the contents of the
package dataset. The stand-alone file is an Excel (\code{.xlsx}) file containing
three sheets: one each containing data (in the same format as above) for
Austin, TX (the data in the package dataset), Fort Worth, TX, and Seattle,
WA.
}
\source{
\code{crimedata} package \url{http://pkgs.lesscrime.info/crimedata/}
}
\usage{
aggravated_assaults
}
\description{
Dataset of aggravated assaults in Austin, TX, in 2019.
}
\keyword{datasets}
