% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicago_aggravated_assaults.R
\docType{data}
\name{chicago_aggravated_assaults}
\alias{chicago_aggravated_assaults}
\title{Aggravated assaults in Chicago, 2010 to 2019}
\format{
Data frame with 148,636 rows and 5 columns:
\describe{
\item{date}{Date-time at which assault occurred}
\item{loc_cat}{Type of location}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
\item{district}{Chicago Police Department District}
}

This dataset is also available as a stand-alone file in comma-separated
values (CSV) format at
\url{https://mpjashby.github.io/crimemappingdata/chicago_aggravated_assaults.csv.gz}

Licence: \url{https://www.chicago.gov/city/en/narr/foia/data_disclaimer.html}

This site provides applications using data that has been modified for use
from its original source, \url{http://www.cityofchicago.org/}, the official
website of the City of Chicago. The City of Chicago makes no claims as to the
content, accuracy, timeliness, or completeness of any of the data provided at
this site. The data provided at this site is subject to change at any time.
It is understood that the data provided at this site is being used at one’s
own risk.
}
\source{
City of Chicago
}
\usage{
chicago_aggravated_assaults
}
\description{
Dataset of aggravated assaults recorded by police in Chicago between 2010 and
2019.
}
\keyword{datasets}
