% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicago_police_districts.R
\docType{data}
\name{chicago_police_districts}
\alias{chicago_police_districts}
\title{Chicago Police Department Districts}
\format{
A simple-features (SF) object containing 25 rows and three columns:
\describe{
\item{Name}{District number}
\item{Description}{District name}
\item{geometry}{SF polygon geometry}
}

This dataset is also available as a stand-alone file in KML format at
\url{https://mpjashby.github.io/crimemappingdata/chicago_police_districts.kml}

Licence: \url{https://www.chicago.gov/city/en/narr/foia/data_disclaimer.html}

This site provides applications using data that has been modified for use
from its original source, \url{http://www.cityofchicago.org/}, the official
website of the City of Chicago. The City of Chicago makes no claims as to the
content, accuracy, timeliness, or completeness of any of the data provided at
this site. The data provided at this site is subject to change at any time.
It is understood that the data provided at this site is being used at one’s
own risk.
}
\source{
City of Chicago
}
\usage{
chicago_police_districts
}
\description{
Boundaries of the geographic districts of the Chicago Police Department, as
of 2018.
}
\keyword{datasets}
