% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cincinnati_burglary.R
\docType{data}
\name{cincinnati_burglary}
\alias{cincinnati_burglary}
\title{Burglaries in Cincinnati, 2016 to 2018}
\format{
Data frame with 8,723 and 5 columns:
\describe{
\item{incident_no}{Unique identifier}
\item{offense_date}{Date on which offence was discovered by victim}
\item{offense}{Offence type}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
}

This dataset is also available as a stand-alone file in comma-separated
values (CSV) format at:
\url{https://mpjashby.github.io/crimemappingdata/cincinnati_burglary.csv.gz}

Licence: Public domain
}
\source{
City of Cincinnati
}
\usage{
cincinnati_burglary
}
\description{
Dataset of aggravated assaults recorded by police in Cincinnati between
January 2016 and December 2018.
}
\keyword{datasets}
