% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czechia_collisions.R
\docType{data}
\name{czechia_collisions}
\alias{czechia_collisions}
\title{Drink/drug-related traffic collisions in Czechia}
\format{
An SF object containing 5,020 rows and 3 columns:
\describe{
\item{id}{Incident number}
\item{date}{Date-time of incident}
\item{geometry}{SF point geometry}
}

This dataset is also available as a stand-alone file in geo-package format at
\url{https://mpjashby.github.io/crimemappingdata/czechia_collisions.gpkg}.

Licence: Bespoke licence, see \url{https://kriminalita.policie.cz/}
}
\source{
Police of the Czech Republic \url{https://kriminalita.policie.cz/}
}
\usage{
czechia_collisions
}
\description{
Dataset of 5,020 vehicle collisions that were related to alcohol or drug use
in Czechia between 2020 and 2022.
}
\keyword{datasets}
