% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czechia_mcycle_thefts.R
\docType{data}
\name{czechia_mcycle_thefts}
\alias{czechia_mcycle_thefts}
\title{Motorcycle thefts in Czechia}
\format{
An SF object containing 9,464 rows and 3 columns:
\describe{
\item{id}{Incident number}
\item{date}{Date-time of incident}
\item{geometry}{SF point geometry}
}

This dataset is also available as a stand-alone file in geo-package format at
\url{https://mpjashby.github.io/crimemappingdata/czechia_mcycle_thefts.gpkg}.

Licence: Bespoke licence, see \url{https://kriminalita.policie.cz/}
}
\source{
Police of the Czech Republic \url{https://kriminalita.policie.cz/}
}
\usage{
czechia_mcycle_thefts
}
\description{
Dataset of 9,464 thefts of two-wheeled motor vehicles in Czechia between 2020
and 2022.
}
\keyword{datasets}
