% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downtown_homicides.R
\docType{data}
\name{downtown_homicides}
\alias{downtown_homicides}
\title{Homicides in downtown Atlanta in 2019}
\format{
A data frame with four rows and four columns:
\describe{
\item{report_number}{A reference number}
\item{label}{A pre-formatted label giving address and date/time}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
}
}
\source{
Atlanta Police Department
}
\usage{
downtown_homicides
}
\description{
Dataset of the four homicides that occurred in the downtown neighbourhood of
Atlanta, GA, in 2019.
}
\keyword{datasets}
