% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glenrose_heights_homicides.R
\docType{data}
\name{glenrose_heights_homicides}
\alias{glenrose_heights_homicides}
\title{Homicides in Glenrose Heights in 2019}
\format{
A data frame with four rows and four columns:
\describe{
\item{report_number}{A reference number}
\item{label}{A pre-formatted label giving address and date/time}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
}

This dataset is also available as a stand-alone file in two formats: in CSV
format at
\url{https://mpjashby.github.io/crimemappingdata/glenrose_heights_homicides.csv}
and in GeoPackage format at
\url{https://mpjashby.github.io/crimemappingdata/glenrose_heights_homicides.gpkg}
}
\source{
Atlanta Police Department
}
\usage{
glenrose_heights_homicides
}
\description{
Dataset of the four homicides that occurred in the Glenrose Heights
neighbourhood of Atlanta, GA, in 2019.
}
\keyword{datasets}
