% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hungerford_shootings.R
\docType{data}
\name{hungerford_shootings}
\alias{hungerford_shootings}
\title{Shootings in the Hungerford Massacre}
\format{
A data frame with 16 rows and 4 columns:
\describe{
\item{victims}{The names of the victims shot at this location, marked with
a dagger symbol (†) if the victim was killed}
\item{order}{The order in which the locations happened in time}
\item{easting}{British National Grid easting co-ordinate}
\item{northing}{British National Grid northing co-ordinate}
}

This dataset is also available as a stand-alone file in CSV format at
\url{https://mpjashby.github.io/crimemappingdata/hungerford_shootings.csv}.

Licence: Open Government Licence
}
\source{
Official report into the massacre
}
\usage{
hungerford_shootings
}
\description{
Dataset of the locations of each shooting during the massacre that took place
in Hungerford, England, on 19 August 1987. Each row represents a location at
which one or more people were shot by the gun-man.
}
\keyword{datasets}
