% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japan_prefectures.R
\docType{data}
\name{japan_prefectures}
\alias{japan_prefectures}
\title{Japanese prefecture boundaries, 2019}
\format{
Data frame of 47 rows and 6 columns:
\describe{
\item{adm0_en}{Top-level administrative unit name (English)}
\item{adm0_ja}{Top-level administrative unit name (Japanese)}
\item{adm0_pcode}{Top-level administrative unit code (English)}
\item{adm1_en}{Prefecture name (English)}
\item{adm1_ja}{Prefecture name (Japanese)}
\item{adm1_pcode}{Prefecture code (English)}
}

This dataset is also available as a stand-alone file in Geopackage
(\code{.gpkg}) format at
\url{https://mpjashby.github.io/crimemappingdata/japan_prefectures.gpkg}.

Licence: Creative Commons Attribution for Intergovernmental Organisations
}
\source{
United Nations Office for the Coordination of Humanitarian Affairs
}
\usage{
japan_prefectures
}
\description{
Dataset of boundaries of Japanese prefectures.
}
\keyword{datasets}
