% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japan_violence_counts.R
\docType{data}
\name{japan_violence_counts}
\alias{japan_violence_counts}
\title{Counts of violent crime in Japan, 2012}
\format{
Data frame of 300 rows and 3 columns:
\describe{
\item{prefecture}{Prefecture name}
\item{measure}{Type of data}
\item{value}{Numeric value}
}

The values of \code{measure} are "number of homicides",
"number of robberies", "number of rapes", "number of violent crimes",
"GDP per capita (¥1000)", "population (thousands)".

This dataset is also available as a stand-alone file in \code{.rds} format at
\url{https://mpjashby.github.io/crimemappingdata/japan_violence_counts.rds}.

Licence: Creative Commons Attribution
}
\source{
Statistics Bureau of Japan
}
\usage{
japan_violence_counts
}
\description{
Dataset of counts of homicides, rapes, robberies and total violent crimes in
each prefecture in Japan in 2012, together with GDP per capita and
population.
}
\keyword{datasets}
