% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kansas_city_frauds.R
\docType{data}
\name{kansas_city_frauds}
\alias{kansas_city_frauds}
\title{Fraud offences in Kansas City in 2015}
\format{
A data frame with 6 rows and 6 columns:
\describe{
\item{uid}{Numeric reference number for each offence}
\item{offense_code}{NIBRS offence code}
\item{offense_type}{NIBRS offence type}
\item{date}{Date-time when the offence occurred}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
}

This dataset is also available as a stand-alone file in CSV format at
\url{https://mpjashby.github.io/crimemappingdata/kansas_city_frauds.csv}.
}
\source{
Kansas City Police Department
}
\usage{
kansas_city_frauds
}
\description{
Dataset of fraud offences recorded by Kansas City Police on 14 February 2015.
}
\keyword{datasets}
