% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancashire_districts.R
\docType{data}
\name{lancashire_districts}
\alias{lancashire_districts}
\title{Local authority districts of Lancashire}
\format{
A simple features (SF) object with 14 rows and 3 columns:
\describe{
\item{objectid}{Government Statistical Service code for each district}
\item{district}{District name}
\item{geometry}{SF polygon geometry}
}

This dataset is also available as a stand-alone file in geoJSON format at
\url{https://mpjashby.github.io/crimemappingdata/lancashire_districts.geojson}.

Licence: Open Government Licence
}
\source{
Office for National Statistics
}
\usage{
lancashire_districts
}
\description{
Dataset of the boundaries of the local authority districts in Lancashire,
England, in 2022.
}
\keyword{datasets}
