% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lancashire_wards.R
\docType{data}
\name{lancashire_wards}
\alias{lancashire_wards}
\title{Local government wards of Lancashire}
\format{
A simple features (SF) object with 271 rows and 3 columns:
\describe{
\item{ward_code}{Government Statistical Service code for each ward}
\item{ward_name}{Ward name}
\item{geometry}{SF polygon geometry}
}

This dataset is also available as a stand-alone file in geopackage format at
\url{https://mpjashby.github.io/crimemappingdata/lancashire_wards.gpkg}.

Licence: Open Government Licence
}
\source{
Office for National Statistics
}
\usage{
lancashire_wards
}
\description{
Dataset of ward boundaries in Lancashire, England, in 2022.
}
\keyword{datasets}
