% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london_attacks.R
\docType{data}
\name{london_attacks}
\alias{london_attacks}
\title{Fatal terrorist attacks in London}
\format{
A data frame containing 8 rows and 4 columns:
\describe{
\item{date}{Attack date}
\item{latitude}{Latitude}
\item{longitude}{Longitude}
\item{description}{Type of attack and number of people killed}
}

A version of this dataset is also available as a stand-alone file at
\url{https://mpjashby.github.io/crimemappingdata/london_attacks.csv}.

Licence: public domain
}
\source{
Compiled by package author
}
\usage{
london_attacks
}
\description{
Dataset of fatal terrorist attacks in London, United Kingdom, from 2010 to
2018.
}
\keyword{datasets}
