% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malaysia_violence_counts.R
\docType{data}
\name{malaysia_violence_counts}
\alias{malaysia_violence_counts}
\title{Violent crime counts in Malaysia, 2006 to 2017}
\format{
Data frame with 840 rows and 5 columns:
\describe{
\item{region}{Region (East or West Malaysia)}
\item{state}{State}
\item{year}{Year}
\item{crime_type}{Type of violent crime}
\item{count}{Count of crimes recorded by police}
}

This dataset is also available as a stand-alone file in \code{.rds}
format at \url{https://mpjashby.github.io/crimemappingdata/malaysia_violence_counts.rds}.

Licence: Creative Commons Attribution
}
\source{
Hakim et al. (2019) \url{https://doi.org/10.1016/j.dib.2019.104449}
}
\usage{
malaysia_violence_counts
}
\description{
Dataset of annual counts of five types of violent crime in each state of
Malaysia for each year from 2006 to 2017.
}
\keyword{datasets}
