% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medellin_metro_lines.R
\docType{data}
\name{medellin_metro_lines}
\alias{medellin_metro_lines}
\title{Metro and cable-car lines in Medellin}
\format{
Simple features (SF) object with 9 rows and 4 columns:
\describe{
\item{linea}{Line letter}
\item{itinerario}{Terminus stations}
\item{sistema}{Metro or Cable}
\item{geometry}{SF line object}
}

This dataset is also available as a stand-alone file in shape-file format
at \url{https://mpjashby.github.io/crimemappingdata/medellin_metro_lines.zip}.
Note the components of the shape file are contained in a zip archive.
}
\source{
Metro de Medellin
}
\usage{
medellin_metro_lines
}
\description{
Routes of metro and cable-car lines in Medellin, Colombia, as of early 2023.
}
\keyword{datasets}
