% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medellin_metro_stns.R
\docType{data}
\name{medellin_metro_stns}
\alias{medellin_metro_stns}
\title{Metro and cable-car stations in Medellin}
\format{
Data frame with 48 rows and 6 columns:
\describe{
\item{linea}{Line letter}
\item{nombre}{Station name}
\item{municipio}{Municipality name}
\item{sistema}{Metro or Cable}
\item{x}{Longitude}
\item{y}{Latitude}
}

This dataset is also available as a stand-alone file in CSV format at
\url{https://mpjashby.github.io/crimemappingdata/medellin_metro_stns.csv}.
}
\source{
Metro de Medellin
}
\usage{
medellin_metro_stns
}
\description{
Locations of metro and cable-car stations in Medellin, Colombia, as of early
2023.
}
\keyword{datasets}
