% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northants_burglary_counts.R
\docType{data}
\name{northants_burglary_counts}
\alias{northants_burglary_counts}
\title{Counts of burglary in Northamptonshire in 2020}
\format{
A data frame with 409 rows and 3 columns:
\describe{
\item{district}{Local authority district}
\item{lsoa}{Lower-layer super output area name}
\item{count}{Number of burglaries}
}

Lower-layer super output areas (LSOAs) are small spatial areas used for
publishing government statistics in the United Kingdom.

This dataset is also available as a stand-alone file in \code{.rds}
format at \url{https://mpjashby.github.io/crimemappingdata/northants_burglary_counts.rds}.

Licence: Open Government Licence
}
\source{
Home Office
}
\usage{
northants_burglary_counts
}
\description{
Dataset of counts of burglaries (residential and commercial combined) in each
lower-layer super output area in Northamptonshire, England, in 2020.
}
\keyword{datasets}
