% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northumbria_asb.R
\docType{data}
\name{northumbria_asb}
\alias{northumbria_asb}
\title{Anti-social behaviour in Northumbria}
\format{
A data frame of 39,706 rows and 5 columns:
\describe{
\item{month}{Date of the first day of the month of the incident}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
\item{location}{Street name or point-of-interest category}
\item{lsoa_code}{Government Statistical Service code for the Lower Layer
Super Output Area in which the incident occurred}
}

This dataset is also available as a stand-alone file in tab-separated values
format at \url{https://mpjashby.github.io/crimemappingdata/northumbria_asb.tab}.

Licence: Open Government Licence
}
\source{
Home Office
}
\usage{
northumbria_asb
}
\description{
Dataset of incidents of anti-social behaviour reported to police in
Northumbria, England, in 2022.
}
\keyword{datasets}
