% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northumbria_districts.R
\docType{data}
\name{northumbria_districts}
\alias{northumbria_districts}
\title{Local authority districts of Northumbria}
\format{
A simple features (SF) object with 5 rows and 3 columns:
\describe{
\item{objectid}{Government Statistical Service code for each district}
\item{district}{District name}
\item{geometry}{SF polygon geometry}
}

This dataset is also available as a stand-alone file in geoJSON format at
\url{https://mpjashby.github.io/crimemappingdata/northumbria_districts.geojson}.

Licence: Open Government Licence
}
\source{
Office for National Statistics
}
\usage{
northumbria_districts
}
\description{
Dataset of the boundaries of the local authority districts in Northumbria,
England, in 2022.
}
\keyword{datasets}
