% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northumbria_ward_pop.R
\docType{data}
\name{northumbria_ward_pop}
\alias{northumbria_ward_pop}
\title{Northumbria ward population}
\format{
A data frame with 177 rows and 3 columns:
\describe{
\item{gss_code}{Government Statistical Service code for each ward}
\item{ward}{Ward name}
\item{population}{Resident population}
}

This dataset is also available as a stand-alone file in Excel (.xlsx) format
at \url{https://mpjashby.github.io/crimemappingdata/northumbria_ward_pop.xlsx}.
}
\source{
Office for National Statistics
}
\usage{
northumbria_ward_pop
}
\description{
Dataset of 2021 resident population of local authority wards in Northumbria,
in mid 2020.
}
\keyword{datasets}
