% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northumbria_wards.R
\docType{data}
\name{northumbria_wards}
\alias{northumbria_wards}
\title{Local government wards of Northumbria}
\format{
A simple features (SF) object with 177 rows and 3 columns:
\describe{
\item{ward_code}{Government Statistical Service code for each ward}
\item{ward_name}{Ward name}
\item{geometry}{SF polygon geometry}
}

This dataset is also available as a stand-alone file in geopackage format at
\url{https://mpjashby.github.io/crimemappingdata/northumbria_wards.gpkg}.

Licence: Open Government Licence
}
\source{
Office for National Statistics
}
\usage{
northumbria_wards
}
\description{
Dataset of ward boundaries in Northumbria, England, in 2022.
}
\keyword{datasets}
