% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nottingham_burglary.R
\docType{data}
\name{nottingham_burglary}
\alias{nottingham_burglary}
\title{Burglaries in Nottingham}
\format{
A data frame of 1,795 rows and 5 columns:
\describe{
\item{month}{Date of the first day of the month of the incident}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
\item{location}{Street name or point-of-interest category}
\item{lsoa_code}{Government Statistical Service code for the Lower Layer
Super Output Area in which the incident occurred}
}

This dataset is also available as a stand-alone file in gzipped CSV format at
\url{https://mpjashby.github.io/crimemappingdata/nottingham_burglary.csv.gz}.

Licence: Open Government Licence
}
\source{
Home Office
}
\usage{
nottingham_burglary
}
\description{
Dataset of burglaries (both residential and non-residential) recorded by
police in Nottingham, England, in 2022.
}
\keyword{datasets}
