% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nottingham_robbery.R
\docType{data}
\name{nottingham_robbery}
\alias{nottingham_robbery}
\title{Robberies in Nottingham}
\format{
A data frame of 555 rows and 5 columns:
\describe{
\item{month}{Date of the first day of the month of the incident}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
\item{location}{Street name or point-of-interest category}
\item{lsoa_code}{Government Statistical Service code for the Lower Layer
Super Output Area in which the incident occurred}
}

This dataset is also available as a stand-alone file in gzipped CSV format at
\url{https://mpjashby.github.io/crimemappingdata/nottingham_robbery.csv.gz}.

Licence: Open Government Licence
}
\source{
Home Office
}
\usage{
nottingham_robbery
}
\description{
Dataset of robberies (both personal and commercial) recorded by police in
Nottingham, England, in 2022.
}
\keyword{datasets}
