% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nottingham_wards.R
\docType{data}
\name{nottingham_wards}
\alias{nottingham_wards}
\title{Local government wards of Nottingham}
\format{
A simple features (SF) object with 20 rows and 3 columns:
\describe{
\item{ward_code}{Government Statistical Service code for each ward}
\item{ward_name}{Ward name}
\item{geometry}{SF polygon geometry}
}

This dataset is also available as a stand-alone file in geopackage format at
\url{https://mpjashby.github.io/crimemappingdata/nottingham_wards.gpkg}.

Licence: Open Government Licence
}
\source{
Office for National Statistics
}
\usage{
nottingham_wards
}
\description{
Dataset of ward boundaries in Nottingham, England, in 2022.
}
\keyword{datasets}
