% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_homicides.R
\docType{data}
\name{nyc_homicides}
\alias{nyc_homicides}
\title{Homicides in New York City, 2016 to 2018}
\format{
Data frame with 904 rows and 4 columns:
\describe{
\item{uid}{Unique numeric identifier}
\item{offense_date}{Offense date}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
}

This dataset is also available as a stand-alone file in comma-separated
values (CSV) format at
\url{https://mpjashby.github.io/crimemappingdata/nyc_homicides.csv}

Licence: NYC Open Data Law
}
\source{
City of New York
}
\usage{
nyc_homicides
}
\description{
Dataset of homicides recorded by police in New York City between 2016 and
2018.
}
\keyword{datasets}
