% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_precincts.R
\docType{data}
\name{nyc_precincts}
\alias{nyc_precincts}
\title{New York City Police precinct boundaries}
\format{
Simple features (SF) object containing 77 rows and 3 columns:
\describe{
\item{OBJECTID}{Row identifier}
\item{Precinct}{Precinct number}
\item{geometry}{SF polygons of the outline of each precinct}
}

This dataset is also available as a stand-alone file in geopackage format
at \url{https://mpjashby.github.io/crimemappingdata/nyc_precincts.gpkg}.

Licence: Creative Commons Attribution \url{https://creativecommons.org/licenses/by/3.0/au/}
}
\source{
City of New York
\url{https://data.cityofnewyork.us/Public-Safety/Police-Precincts/78dh-3ptz}
}
\usage{
nyc_precincts
}
\description{
Dataset of the boundaries of each precinct of the New York City Police
Department (NYPD), 2022.
}
\keyword{datasets}
