% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qld_police_divisions.R
\docType{data}
\name{qld_police_divisions}
\alias{qld_police_divisions}
\title{Queensland Police Service division boundaries}
\format{
Simple features (SF) object containing 331 rows and 2 columns:
\describe{
\item{division}{Police division name}
\item{geometry}{SF polygons of the outline of each division}
}

This dataset is also available as a stand-alone file in geopackage format
at \url{https://mpjashby.github.io/crimemappingdata/qld_police_divisions.gpkg}.

Licence: Creative Commons Attribution \url{https://creativecommons.org/licenses/by/3.0/au/}
}
\source{
Queensland Government
\url{https://www.data.qld.gov.au/dataset/qps-divisions/resource/fa6a7917-43de-4036-a704-25f545c24093}
}
\usage{
qld_police_divisions
}
\description{
Dataset of the boundaries of each division of the Queensland Police Service,
2022.
}
\keyword{datasets}
