% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qld_population.R
\docType{data}
\name{qld_population}
\alias{qld_population}
\title{Population of Queensland Police Service divisions}
\format{
Data frame containing 334 rows and 2 columns:
\describe{
\item{police_division}{Police division name}
\item{population}{Number of usual residents}
}

This dataset is also available as a stand-alone file in gzipped CSV format
at \url{https://mpjashby.github.io/crimemappingdata/qld_population.csv.gz}.
}
\source{
Queensland Government
}
\usage{
qld_population
}
\description{
Dataset of resident population of each Queensland Police Service division,
2016.
}
\keyword{datasets}
