% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qld_stalking.R
\docType{data}
\name{qld_stalking}
\alias{qld_stalking}
\title{Stalking in Queensland from 2001 to 2022}
\format{
A data frame containing 7,304 rows and 3 columns:
\describe{
\item{division}{Name of police visions}
\item{year}{Year}
\item{stalking}{Count of recorded stalking offences}
}

This dataset is also available as a stand-alone file in Excel (.xlsx) format
at \url{https://mpjashby.github.io/crimemappingdata/qld_stalking.xlsx}.

Licence: Creative Commons Attribution \url{https://creativecommons.org/licenses/by/4.0/}
}
\source{
Queensland Police Service
\url{https://www.police.qld.gov.au/maps-and-statistics}
}
\usage{
qld_stalking
}
\description{
Dataset of annual counts of stalking offences in each police division in
Queensland, Australia, from 2001 to 2022.
}
\keyword{datasets}
