% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/san_francisco_robbery.R
\docType{data}
\name{san_francisco_robbery}
\alias{san_francisco_robbery}
\title{Personal robberies in San Francisco in 2019}
\format{
A data frame containing 951 rows and 5 columns:
\describe{
\item{uid}{Numeric reference number for each crime}
\item{offense_type}{Offence type (always "personal robbery")}
\item{date_time}{Date-time at which offence occurred}
\item{longitude}{Longitude}
\item{latitude}{Latitude}
}

This dataset is also available as a stand-alone file in CSV format at
\url{https://mpjashby.github.io/crimemappingdata/san_francisco_robbery.csv}.

Licence: Open Data Commons Public Domain Dedication and License \url{https://opendatacommons.org/licenses/pddl/1.0/}
}
\source{
San Francisco Police Department
}
\usage{
san_francisco_robbery
}
\description{
Dataset of personal robberies in San Francisco in 2019.
}
\keyword{datasets}
