% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sao_paulo_homicides.R
\docType{data}
\name{sao_paulo_homicides}
\alias{sao_paulo_homicides}
\title{Homicides in the region of Sao Paulo}
\format{
A data frame containing 6,214 rows and 7 columns:
\describe{
\item{year}{Year}
\item{date}{Homicide date}
\item{municipality}{Municipality}
\item{location_type}{Type of location}
\item{victim_sex}{Victim sex}
\item{latitude}{Latitude}
\item{longitude}{Longitude}
}

A version of this dataset is also available as a stand-alone file at
\url{https://mpjashby.github.io/crimemappingdata/sao_paulo_homicides.xlsx}.

Licence: public domain
}
\source{
Compiled by package author
}
\usage{
sao_paulo_homicides
}
\description{
Dataset of homicides recorded by police in the immediate region of Sao Paulo
from 2017 to 2022.
}
\keyword{datasets}
