% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sao_paulo_muni.R
\docType{data}
\name{sao_paulo_muni}
\alias{sao_paulo_muni}
\title{Sao Paolo municipality boundaries}
\format{
A data frame containing 60 rows and 3 columns:
\describe{
\item{code_muni}{Statistical code for municipality}
\item{name_muni}{Municipality name}
\item{geom}{Feature geometry}
}

A version of this dataset is also available as a stand-alone file at
\url{https://mpjashby.github.io/crimemappingdata/sao_paulo_muni.geojson}.

Licence: public domain
}
\source{
Compiled by package author
}
\usage{
sao_paulo_muni
}
\description{
Dataset of municipality boundaries in the immediate region of Sao Paulo, 2010.
}
\keyword{datasets}
