% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uttar_pradesh_districts.R
\docType{data}
\name{uttar_pradesh_districts}
\alias{uttar_pradesh_districts}
\title{Districts of Uttar Pradesh}
\format{
Simple features (SF) object containing 75 rows and 3 columns:
\describe{
\item{state}{State name}
\item{district_name}{District name}
\item{geometry}{SF geometry representation of the district boundary}
}

This dataset is also available as a stand-alone file in geopackage format at
\url{https://mpjashby.github.io/crimemappingdata/uttar_pradesh_districts.gpkg}.

Licence: MIT
}
\source{
Hindustan Times
\url{https://github.com/HindustanTimesLabs/shapefiles}
}
\usage{
uttar_pradesh_districts
}
\description{
Spatial dataset of the boundaries of administrative districts in the Indian
state of uttar Pradesh.
}
\keyword{datasets}
