% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uttar_pradesh_murders.R
\docType{data}
\name{uttar_pradesh_murders}
\alias{uttar_pradesh_murders}
\title{Murders in Uttar Pradesh}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 75 rows and 2 columns.
}
\source{
Government of India
\url{https://data.gov.in/resources/district-wise-crime-under-various-sections-indian-penal-code-ipc-crimes-during-2014}
}
\usage{
uttar_pradesh_murders
}
\description{
Dataset containing counts of murders in each district in the Indian state of
Uttar Pradesh in 2014.
}
\details{
Data frame containing 75 rows and 2 columns:
\describe{
\item{district}{District name}
\item{murder}{Count of murders}
}

This dataset is also available as a stand-alone file in CSV format at
\url{https://mpjashby.github.io/crimemappingdata/uttar_pradesh_murders.csv}.

Licence: Government Open Data License \url{https://data.gov.in/government-open-data-license-india}
}
\keyword{datasets}
