% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmx_alcaldias.R
\docType{data}
\name{cdmx_alcaldias}
\alias{cdmx_alcaldias}
\title{Alcaldías (boroughs) of Mexico City}
\format{
Spatial dataset containing 16 rows and 3 columns:
\describe{
\item{municip}{Row number}
\item{nomgeo}{Name}
\item{geometry}{SF polygon geometry}
}

This dataset is also available as a stand-alone file in geopackage
format at \url{https://mpjashby.github.io/crimemappingdata/cdmx_alcaldias.gpkg}.

Licence: Creative Commons Attribution \url{https://creativecommons.org/licenses/by/4.0/}
}
\source{
Government of Mexico City
}
\usage{
cdmx_alcaldias
}
\description{
Spatial dataset of the boundaries of the alcaldías (boroughs) of Mexico City.
}
\keyword{datasets}
