% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmx_car_jacking.R
\docType{data}
\name{cdmx_car_jacking}
\alias{cdmx_car_jacking}
\title{Car jackings in Mexico City}
\format{
Simple features (SF) object containing 2,811 rows and 4 columns:
\describe{
\item{fecha_hechos}{Date and time of the offence}
\item{longitud}{Longitude}
\item{latitud}{Latitude}
\item{geometry}{SF point geometry}
}

This dataset is also available as a stand-alone file in geopackage format
at \url{https://mpjashby.github.io/crimemappingdata/cdmx_car_jacking.gpkg}.

Licence: Creative Commons Attribution \url{https://creativecommons.org/licenses/by/4.0/}
}
\source{
Mexico City Attorney General’s Office
}
\usage{
cdmx_car_jacking
}
\description{
Dataset of car jackings (\emph{"robo de vehiculo de servicio particular con violencia"}) in Mexico City in 2019.
}
\keyword{datasets}
