% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempe_opioid_calls.R
\docType{data}
\name{tempe_opioid_calls}
\alias{tempe_opioid_calls}
\title{Calls to opioid overdoses in Tempe, Arizona}
\format{
A data frame containing 3,813 rows and 5 columns:
\describe{
\item{year}{Year of incident}
\item{incident_date}{Date of incident}
\item{narcan_given}{Was naloxone given to the patient?}
\item{latitude}{Latitude}
\item{longitude}{Longitude}
}

A version of this dataset is also available as a stand-alone file at
\url{https://mpjashby.github.io/crimemappingdata/tempe_opioid_calls.csv}.

Licence: Creative Commons Attribution (CC-BY)
}
\source{
Open Tempe
}
\usage{
tempe_opioid_calls
}
\description{
Dataset of calls to emergency medical services in Tempe for suspected opioid
overdoses.
}
\keyword{datasets}
