% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uttar_pradesh_population.R
\docType{data}
\name{uttar_pradesh_population}
\alias{uttar_pradesh_population}
\title{Population of districts in Uttar Pradesh}
\format{
Data frame containing 75 rows and 7 columns:
\describe{
\item{code}{Abbreviation code}
\item{district}{District name}
\item{headquarters}{Administrative centre for the district}
\item{division}{Division in which the district is located}
\item{population}{Usual residents, 2011}
\item{area}{Area, square kilometres}
\item{density_km2}{Population density, people per square kilometre}
}

This dataset is also available as a stand-alone file in CSV format
at \url{https://mpjashby.github.io/crimemappingdata/uttar_pradesh_population.csv}.

Licence: Creative Commons Attribution \url{https://creativecommons.org/licenses/by/3.0/}
}
\source{
Registrar General and Census Commissioner of India
}
\usage{
uttar_pradesh_population
}
\description{
Dataset of resident population of each administrative district in Uttar
Pradesh, India, in 2011.
}
\keyword{datasets}
